+++ Assets/Architecture/Domain/Calculations/UnsignedOperator.cs
@@ -1,0 +1,9 @@
+﻿namespace Calc.Domain.Calculations
+{
+    using Functional;
+
+    public abstract record UnsignedBinaryOperator
+    {
+        public abstract Result<UnsignedOperand> Calculate(UnsignedOperand first, UnsignedOperand second);
+    }
+}