+++ Assets/Architecture/Domain/Functional/Outcome/Success.cs
@@ -1,0 +1,7 @@
+﻿using System.Runtime.InteropServices;
+
+namespace Calc.Domain.Functional.Outcome
+{
+    [StructLayout(LayoutKind.Sequential, Size = 1)]
+    public readonly struct Success { }
+}